+++ ClasseAluno/Program.cs
@@ -5,7 +5,7 @@ internal class Program
     {
         Aluno a1 = new Aluno();
         System.Console.Write("Digite o RA do aluno: ");
-        a1.ra = Convert.ToInt32(Console.ReadLine());
+        a1.ra = Convert.ToInt64(Console.ReadLine());
         System.Console.Write("Digite o nome do aluno: ");
         a1.nome = Console.ReadLine();
         System.Console.Write("Digite a nota da P1: ");
@@ -16,7 +16,7 @@ internal class Program
 
         Aluno a2 = new Aluno();
         System.Console.Write("Digite o RA do aluno: ");
-        a2.ra = Convert.ToInt32(Console.ReadLine());
+        a2.ra = Convert.ToInt64(Console.ReadLine());
         System.Console.Write("Digite o nome do aluno: ");
         a2.nome = Console.ReadLine();
         System.Console.Write("Digite a nota da P1: ");
